+++ Assets/Resources/ScriptableObject/GameDataOriginal.asset
@@ -12,6 +12,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: dac803bb941873b4fafc0c839fc31480, type: 3}
   m_Name: GameDataOriginal
   m_EditorClassIdentifier: 
+  Incarna: {fileID: 11400000, guid: 8f107784ecc1ae64c827a89a87b95263, type: 2}
   Money: 0
   DarkEssence: 0
   DeckUnits:
