+++ Assets/Resources/ScriptableObject/Incarna.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 127061bd89cf03b48ad35386e18eff48
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
