+++ Assets/Resources/ScriptableObject/Incarna/Incarna.asset
@@ -1,0 +1,21 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: bd7ba0f17e0a1ac42bfb7baaf8e8e99b, type: 3}
+  m_Name: Incarna
+  m_EditorClassIdentifier: 
+  Name: "\uD654\uC2E0"
+  Sprite: {fileID: 0}
+  PlayerSkillList:
+  - {fileID: 11400000, guid: a8c7559f612cb8a429662984cdad7087, type: 2}
+  - {fileID: 11400000, guid: 9619319320dfc73468abf666a2ca7034, type: 2}
+  - {fileID: 11400000, guid: b54a8847ace00584ca9174d138240985, type: 2}
+  - {fileID: 11400000, guid: 83cab6ffc5c178a4386cab417f65780a, type: 2}
