+++ Assets/Resources/ScriptableObject/Incarna/Incarna.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8f107784ecc1ae64c827a89a87b95263
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
