+++ Assets/Resources/ScriptableObject/PlayerSkill.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1a00cce653330b4419d5a4d4ba2674f4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
