+++ Assets/Resources/ScriptableObject/PlayerSkill/PlayerSkill_Bounce.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 83cab6ffc5c178a4386cab417f65780a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
