+++ Assets/Resources/ScriptableObject/PlayerSkill/PlayerSkill_Damage.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9619319320dfc73468abf666a2ca7034
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
