+++ Assets/Resources/ScriptableObject/PlayerSkill/PlayerSkill_Heal.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b54a8847ace00584ca9174d138240985
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
