+++ Assets/Resources/ScriptableObject/PlayerSkill/PlayerSkill_Whisper.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a8c7559f612cb8a429662984cdad7087
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
