+++ Assets/Script/Data/GameData.cs
@@ -5,9 +5,10 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "GameData", menuName = "Scriptable Object/Data")]
 public class GameData : ScriptableObject
 {
+    public Incarna Incarna;
     public int Money;
     public int DarkEssence;
-    public List<DeckUnit> DeckUnits = new List<DeckUnit>();
+    public List<DeckUnit> DeckUnits = new();
     public bool isVisitUpgrade = false;
     public bool isVisitStigma = false;
 }
