+++ Assets/Script/ScriptableObject/Incarna.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bd7ba0f17e0a1ac42bfb7baaf8e8e99b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 200
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
