+++ Assets/Script/ScriptableObject/PlayerSkill/PlayerSkill.cs
@@ -1,8 +1,9 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+[CreateAssetMenu(fileName = "PlayerSkill", menuName = "Scriptable Object/PlayerSkill")]
 
-public abstract class PlayerSkill : MonoBehaviour
+public abstract class PlayerSkill : ScriptableObject
 {
     public abstract string GetName();
     public abstract int GetManaCost();
