+++ Assets/Script/ScriptableObject/PlayerSkill/PlayerSkill_Heal.cs
@@ -1,13 +1,14 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+[CreateAssetMenu(fileName = "PlayerSkillHeal", menuName = "Scriptable Object/PlayerSkillHeal")]
 
-public class PlayerSkillBounce : PlayerSkill
+public class PlayerSkill_Heal : PlayerSkill
 {
-    private string name = "Bounce";
+    private string name = "Heal";
     private int manaCost = 20;
     private int darkEssence = 0;
-    private string description = "20 마나를 지불하고 원하는 아군 유닛을 손으로 가져옵니다.";
+    private string description = "20 마나를 지불하고 원하는 유닛의 체력을 20 회복합니다.";
 
     public override int GetDarkEssenceCost() => darkEssence;
     public override int GetManaCost() => manaCost;
@@ -16,11 +17,11 @@ public class PlayerSkillBounce : PlayerSkill
 
     public override void CancelSelect()
     {
-        BattleManager.PlayerSkillController.FriendlyTargetPlayerSkillReady(FieldColorType.none);
+        BattleManager.PlayerSkillController.UnitTargetPlayerSkillReady(FieldColorType.none);
     }
 
     public override void OnSelect()
     {
-        BattleManager.PlayerSkillController.FriendlyTargetPlayerSkillReady(FieldColorType.PlayerSkillBounce);
+        BattleManager.PlayerSkillController.UnitTargetPlayerSkillReady(FieldColorType.PlayerSkillHeal);
     }
 }
