+++ Assets/Script/ScriptableObject/PlayerSkill/PlayerSkill_Whisper.cs
@@ -1,20 +1,19 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+[CreateAssetMenu(fileName = "PlayerSkillWhisper", menuName = "Scriptable Object/PlayerSkillWhisper")]
 
-public class PlayerSkillDamage : PlayerSkill
+public class PlayerSkill_Whisper : PlayerSkill
 {
-    private string name = "Damage";
+    private string name = "Whisper";
     private int manaCost = 20;
-    private int darkEssence = 0;
-    private string description = "20 마나를 지불하고 원하는 적 유닛에게 20 대미지를 줍니다.";
-
+    private int darkEssence = 1;
+    private string description = "20 마나와 1 검은 정수를 지불하고 원하는 적 유닛에게 타락도 1을 부여합니다.";
 
     public override int GetDarkEssenceCost() => darkEssence;
     public override int GetManaCost() => manaCost;
     public override string GetName() => name;
     public override string GetDescription() => description;
-
 
     public override void CancelSelect()
     {
@@ -23,6 +22,6 @@ public class PlayerSkillDamage : PlayerSkill
 
     public override void OnSelect()
     {
-        BattleManager.PlayerSkillController.EnemyTargetPlayerSkillReady(FieldColorType.PlayerSkillDamage);
+        BattleManager.PlayerSkillController.EnemyTargetPlayerSkillReady(FieldColorType.PlayerSkillWhisper);
     }
 }
